+++ wsi_service/slide_manager.py
@@ -1,5 +1,4 @@
 import asyncio
-import atexit
 import os
 
 import aiohttp
@@ -21,20 +20,24 @@ class SlideManager:
         self.mapper_address = mapper_address
         self.data_dir = data_dir
         self.timeout = timeout
+        self.storage_mapper_cache = {}
         self.opened_slide_storages = {}
         self.lock = asyncio.Lock()
         self.storage_locks = {}
         self.event_loop = asyncio.get_event_loop()
-        atexit.register(self.close)
 
     def close(self):
         for storage_address in list(self.opened_slide_storages.keys()):
             self.opened_slide_storages[storage_address].timer.cancel()
-            self._close_slide(storage_address)
+            self._sync_close_slide(storage_address)
 
     async def get_slide(self, slide_id):
-        main_storage_address = await self._get_slide_main_storage_address(slide_id)
-        storage_address = os.path.join(self.data_dir, main_storage_address["address"])
+        if slide_id in self.storage_mapper_cache:
+            storage_address = self.storage_mapper_cache[slide_id]
+        else:
+            main_storage_address = await self._get_slide_main_storage_address(slide_id)
+            storage_address = os.path.join(self.data_dir, main_storage_address["address"])
+            self.storage_mapper_cache[slide_id] = storage_address
         logger.debug("Get slide %s at storage address: %s", slide_id, storage_address)
 
         await self._set_storage_lock(storage_address)
@@ -63,15 +66,14 @@ class SlideManager:
                 self.storage_locks[storage_address] = asyncio.Lock()
 
     async def _open_slide(self, storage_address):
-        # we set a dummy id here, that will be overwriteen before returning the slide handle
-        slide = await load_slide(storage_address, "dummy-id")
+        slide = await load_slide(storage_address)
         self.opened_slide_storages[storage_address] = ExpiringSlide(slide)
 
     def _reset_slide_expiration(self, storage_address):
         expiring_slide = self.opened_slide_storages[storage_address]
         if expiring_slide.timer is not None:
             expiring_slide.timer.cancel()
-        expiring_slide.timer = self.event_loop.call_later(self.timeout, self._close_slide, storage_address)
+        expiring_slide.timer = self.event_loop.call_later(self.timeout, self._sync_close_slide, storage_address)
         logger.debug("Set expiration timer for storage address (%s): %s", storage_address, self.timeout)
 
     async def _get_slide_main_storage_address(self, slide_id):
@@ -92,8 +94,11 @@ class SlideManager:
                 return storage_address
         return slide["storage_addresses"][0]
 
-    def _close_slide(self, storage_address):
+    def _sync_close_slide(self, storage_address):
+        asyncio.create_task(self._close_slide(storage_address))
+
+    async def _close_slide(self, storage_address):
         if storage_address in self.opened_slide_storages:
-            self.opened_slide_storages[storage_address].slide.close()
+            await self.opened_slide_storages[storage_address].slide.close()
             del self.opened_slide_storages[storage_address]
             logger.debug("Closed slide with storage address: %s", storage_address)