+++ wsi_service_base_plugins/tifffile/wsi_service_plugin_tifffile/__init__.py
@@ -11,5 +11,5 @@ supported_file_extensions = [
 ]
 
 
-async def open(filepath, slide_id=0):
-    return Slide(filepath, slide_id)
+async def open(filepath):
+    return await Slide.create(filepath)